# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "fox-admin 后台管理接口文档",
        "title": "fox-admin API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/system/auth/login": {
            "post": {
                "description": "使用后台账号密码登录并签发 access token 和 refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备 ID",
                        "name": "X-Device-ID",
                        "in": "header"
                    },
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/system/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "吊销当前 access token 对应的登录会话",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统认证"
                ],
                "summary": "退出登录",
                "responses": {
                    "200": {
                        "description": "退出成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/system/auth/refresh": {
            "post": {
                "description": "使用请求头中的 refresh token 轮换登录凭证",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统认证"
                ],
                "summary": "刷新登录凭证",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Refresh Token",
                        "name": "X-Refresh-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/system/auth/routers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查询当前登录用户通过角色获得的 Arco Pro 动态路由树",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统认证"
                ],
                "summary": "查询当前用户动态路由",
                "responses": {
                    "200": {
                        "description": "动态路由树",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/system/auth/user-info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查询当前登录用户、角色编码和权限标识",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统认证"
                ],
                "summary": "查询当前用户信息",
                "responses": {
                    "200": {
                        "description": "当前用户信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.LoginReq": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}